feat: implement BitcoinInfoBloc
The bloc retrieves basic information about the bitcoin
network and deamon health

+++ lib/bitcoind/models/bitcoin_network_info.dart
@@ -126,7 +126,7 @@ class LocalAddress {
   LocalAddress.fromJson(Map<String, dynamic> json) {
     address = json['address'];
     port = json['port'];
-    score = json['score'];
+    score = json['score'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
